PayPal Express integration: Fix received_at
The received_at field was being parsed in such a way that the timezone
was being ignored; the most difficult part of fixing it was getting
something that worked on 1.8 as well as 1.9.

The only potential impact of this would be on a consumer of the method
that assumed a particular timezone; previously the timezone was the
callers, but now it is always UTC.

Fixes #542.