Bogus gateway: Add authorization to response
Most (if not all) gateways add 'authorization' to response of purchase
transaction. It allows developer to use returned 'authorization' as a
payment source for subsequent transactions in some
gateways (PayJunction for example).

To test this behavior properly it makes sense to have BogusGateway to
return 'authorization' in purchase response too.

Closes #677.