Barclays ePDQ: Handle incorrectly encoded response
When an order contains non-standard characters (normally in the
customer's address and is normally accented characters) and is submitted
to Barclays, they will send back an XML response flagged as UTF-8 but is
actually ISO-8859-1. This causes parsers like REXML to raise an
exception while parsing what it thinks is UTF-8.

Barclays has acknowledged this issue is on their end and said they have
no plans on fixing it.

Fixed by forcing the encoding of all responses to ISO-8859-1.

Closes #588, #683.