Litle: Allow easier access to the response code
Prior to this commit, the response code could show up within a number of
different response param keys, depending on whether it was an authorize,
a purchase, etc.

Now, we have the 'response_code' key in the Response params allowing
clients to get the code in the same way regardless of the operation.
In addition, Active Merchant uses the response_code key for a number of
other gateways.

Closes #905.