Braintree Blue: Allow unstoring just a credit card
Unfortunately the Braintree implementation of store/unstore/update was
originally made very customer-centric where ActiveMerchant is credit
card centric. This means that #unstore expects a customer id rather than
a credit card token. To at least allow unstoring a card only, this
allows passing a credit card token for removal; the customer id must be
passed as nil for it to work, though.